requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-core-tracing-opentelemetry==1.0.0b11
azure-functions==1.19.0
azure-monitor-opentelemetry==1.4.1
azure-monitor-opentelemetry-exporter==1.0.0b25
botbuilder==0.0.1
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
fixedint==0.1.6
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.0
isodate==0.6.1
jsonpickle==1.4.2
msal==1.28.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
openai==1.26.0
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-dbapi==0.45b0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-openai==0.17.1
opentelemetry-instrumentation-psycopg2==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-urllib==0.45b0
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-resource-detector-azure==0.1.4
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-semantic-conventions-ai==0.2.0
opentelemetry-util-http==0.45b0
packaging==24.0
psutil==5.9.8
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
ujson==5.9.0
urllib3==1.26.18
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
